endpoint: move configuration functions to `config.go`
Signed-off by: Ian Vernon <ian@cilium.io>
